Basic Entity Models
I will upload the Python scraper scaffold as soon as I manage to test it with the new DB.  I'd also like to add some setup instructions as well.  It's not too tricky, but I figure its best to be as detailed as possible to avoid any build issues.  Those are by far the most frustrating issues.