Add comments to entity models + mandate frequency
We did briefly discuss the idea of there being 3 separate "base times" (small, medium, large) that could be used in place of a custom frequency, but for now, I think we should just mandate frequency for the sake of simplicity.  If we decide to implement that idea later, it shouldn't be too hard.

It will be easier to make a mandated field optional than make an optional field mandated.